完成消息通知

+++ common/src/main/java/com/liucj/common/Application.java
@@ -1,14 +1,24 @@
-package com.liucj.factory;
+package com.liucj.common;
 
 import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
 import android.os.SystemClock;
+import android.widget.Toast;
+
+import androidx.annotation.StringRes;
+
+import net.qiujuer.genius.kit.handler.Run;
+import net.qiujuer.genius.kit.handler.runable.Action;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * @author qiujuer Email:qiujuer@live.cn
+ * @version 1.0.0
+ */
 public class Application extends android.app.Application {
     private static Application instance;
     private List<Activity> activities = new ArrayList<>();
@@ -17,7 +27,8 @@ public class Application extends android.app.Application {
     public void onCreate() {
         super.onCreate();
         instance = this;
-        Factory.initDBFlow();
+
+
         registerActivityLifecycleCallbacks(new ActivityLifecycleCallbacks() {
             @Override
             public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
@@ -57,7 +68,7 @@ public class Application extends android.app.Application {
     }
 
     // 退出所有
-    public void finishAll(){
+    public void finishAll() {
         for (Activity activity : activities) {
             activity.finish();
         }
@@ -65,7 +76,7 @@ public class Application extends android.app.Application {
         showAccountView(this);
     }
 
-    protected void showAccountView(Context context){
+    protected void showAccountView(Context context) {
 
     }
 
@@ -126,14 +137,45 @@ public class Application extends android.app.Application {
         File[] files = dir.listFiles();
         if (files != null && files.length > 0) {
             for (File file : files) {
+                boolean isTempFile = file.getName().lastIndexOf("tmp.mp3") != -1;
+                if (isTmp != isTempFile) {
+                    continue;
+                }
                 //noinspection ResultOfMethodCallIgnored
                 file.delete();
             }
         }
 
         // aar
-        File path = new File(getCacheDirFile(), isTmp ? "tmp.mp3" : SystemClock.uptimeMillis() + ".mp3");
+        File path = new File(dir, isTmp ? "tmp.mp3" : SystemClock.uptimeMillis() + ".mp3");
         return path.getAbsoluteFile();
+    }
+
+    /**
+     * 显示一个Toast
+     *
+     * @param msg 字符串
+     */
+    public static void showToast(final String msg) {
+        // Toast 只能在主线程中显示，所有需要进行线程转换，
+        // 保证一定是在主线程进行的show操作
+        Run.onUiAsync(new Action() {
+            @Override
+            public void call() {
+                // 这里进行回调的时候一定就是主线程状态了
+                Toast.makeText(instance, msg, Toast.LENGTH_SHORT).show();
+            }
+        });
+
+    }
+
+    /**
+     * 显示一个Toast
+     *
+     * @param msgId 传递的是字符串的资源
+     */
+    public static void showToast(@StringRes int msgId) {
+        showToast(instance.getString(msgId));
     }
 
 }
